chore: remove unreachable client components and their tests
A reachability walk from each module's declared entry points
(client/index.js, server/index.js, settingsComponent, sotuWidgets)
found 42 client files that cannot be reached from any menu. Some
were replaced and never removed, some were never wired up.

Removes 34 .vue components and 8 supporting composables/utils,
plus the 8 test files that only exercised deleted code. Two test
files covered a mix of live and dead components, so only the dead
describe blocks were removed; health-components.test.js is renamed
to big-rock-expanded-row.test.js to match what it still tests.

Notable removals include the plan health dashboard and its table,
the abandoned releases copies of the commitment tracking and
post-release defects reports (OKR hub owns the live versions), and
two of the three FeatureTable variants.

Deliberately left in place: two server-side files and the generated
system-health quality report registry, which have a different risk
profile and need separate review.

No behaviour change - none of these files were reachable at runtime.
Verified with a re-run of the reachability walk (no cascade), the
full test suite (4396 passing) and eslint.
