+++ modules/releases/__tests__/client/plan/components.test.js
@@ -1,159 +1,8 @@
 import { describe, it, expect } from 'vitest'
 import { mount } from '@vue/test-utils'
-import StatusBadge from '../../../client/plan/components/StatusBadge.vue'
-import SummaryCards from '../../../client/plan/components/SummaryCards.vue'
 import BigRockRow from '../../../client/plan/components/BigRockRow.vue'
 import BigRocksTable from '../../../client/plan/components/BigRocksTable.vue'
 import FilterBar from '../../../client/plan/components/FilterBar.vue'
-
-describe('StatusBadge', () => {
-  it('renders status text', () => {
-    const wrapper = mount(StatusBadge, { props: { status: 'In Progress' } })
-    expect(wrapper.text()).toBe('In Progress')
-  })
-
-  it('applies color class for known statuses', () => {
-    const wrapper = mount(StatusBadge, { props: { status: 'In Progress' } })
-    expect(wrapper.find('span').classes()).toEqual(
-      expect.arrayContaining([expect.stringContaining('blue')])
-    )
-  })
-
-  it('applies default color for unknown statuses', () => {
-    const wrapper = mount(StatusBadge, { props: { status: 'Unknown Status' } })
-    expect(wrapper.find('span').classes()).toEqual(
-      expect.arrayContaining([expect.stringContaining('gray')])
-    )
-  })
-})
-
-describe('SummaryCards', () => {
-  const summary = {
-    totalFeatures: 42,
-    totalRfes: 10,
-    totalBigRocks: 5,
-    rocksWithData: 3,
-    tier1: { features: 19, rfes: 5 },
-    tier2: { features: 15, rfes: 5 },
-    tier3: { features: 8, rfes: 0 }
-  }
-
-  it('renders all three summary cards', () => {
-    const wrapper = mount(SummaryCards, { props: { summary, tier1HealthSummary: null, releaseDistribution: null } })
-    expect(wrapper.text()).toContain('Feature Summary')
-    expect(wrapper.text()).toContain('Big Rock Health')
-    expect(wrapper.text()).toContain('Milestone Coverage')
-    // Should have exactly 3 card root elements in the grid
-    const cards = wrapper.findAll('.p-4.rounded-lg')
-    expect(cards).toHaveLength(3)
-  })
-
-  it('shows correct feature and RFE counts', () => {
-    const wrapper = mount(SummaryCards, { props: { summary, tier1HealthSummary: null, releaseDistribution: null } })
-    // Total counts in Feature Summary subtitle
-    expect(wrapper.text()).toContain('42')
-    expect(wrapper.text()).toContain('10')
-    // Table element exists with aria-label
-    const table = wrapper.find('table[aria-label]')
-    expect(table.exists()).toBe(true)
-  })
-
-  it('renders nothing when summary is null', () => {
-    const wrapper = mount(SummaryCards, { props: { summary: null } })
-    expect(wrapper.text()).toBe('')
-  })
-
-  it('shows per-tier breakdown table in Feature Summary card', () => {
-    const wrapper = mount(SummaryCards, {
-      props: { summary, tier1HealthSummary: null, releaseDistribution: null }
-    })
-    // Table exists with accessible label
-    const table = wrapper.find('table[aria-label]')
-    expect(table.exists()).toBe(true)
-    // Header row
-    expect(table.text()).toContain('Features')
-    expect(table.text()).toContain('RFEs')
-    // Tier rows
-    expect(table.text()).toContain('Tier 1')
-    expect(table.text()).toContain('Tier 2')
-    expect(table.text()).toContain('Tier 3')
-  })
-
-  it('shows health bars when tier1HealthSummary is provided', () => {
-    const tier1HealthSummary = { byRisk: { green: 8, yellow: 4, red: 2 } }
-    const wrapper = mount(SummaryCards, {
-      props: { summary, tier1HealthSummary, releaseDistribution: null }
-    })
-    expect(wrapper.text()).toContain('Big Rock Health')
-    expect(wrapper.text()).toContain('features tracked')
-    // Health section with aria-label
-    const healthSection = wrapper.find('[aria-label*="on track"]')
-    expect(healthSection.exists()).toBe(true)
-    expect(healthSection.attributes('aria-label')).toContain('8 on track')
-    expect(healthSection.attributes('aria-label')).toContain('4 at risk')
-    expect(healthSection.attributes('aria-label')).toContain('2 off track')
-    // Labels and counts visible
-    expect(healthSection.text()).toContain('On track')
-    expect(healthSection.text()).toContain('At risk')
-    expect(healthSection.text()).toContain('Off track')
-    expect(healthSection.text()).toContain('8')
-    expect(healthSection.text()).toContain('4')
-    expect(healthSection.text()).toContain('2')
-  })
-
-  it('shows features tracked count in Outcome Health card', () => {
-    const wrapper = mount(SummaryCards, {
-      props: { summary, tier1HealthSummary: null, releaseDistribution: null }
-    })
-    // Shows feature count from summary.tier1.features
-    expect(wrapper.text()).toContain('19')
-    expect(wrapper.text()).toContain('features tracked')
-  })
-
-  it('shows loading state when tier1HealthSummary is null', () => {
-    const wrapper = mount(SummaryCards, {
-      props: { summary, tier1HealthSummary: null, releaseDistribution: null }
-    })
-    expect(wrapper.find('[aria-label="Loading health data"]').exists()).toBe(true)
-    expect(wrapper.findAll('.animate-pulse').length).toBeGreaterThanOrEqual(3)
-    // No health bars rendered
-    expect(wrapper.find('[aria-label*="on track"]').exists()).toBe(false)
-  })
-
-  it('renders release distribution with doughnut chart and legend', () => {
-    const releaseDistribution = [
-      { label: 'EA1', count: 3, pct: 30, barColor: 'bg-blue-500' },
-      { label: 'GA', count: 4, pct: 40, barColor: 'bg-emerald-500' },
-      { label: '--', count: 3, pct: 30, barColor: 'bg-gray-400' }
-    ]
-    const wrapper = mount(SummaryCards, {
-      props: { summary, releaseDistribution }
-    })
-    // Legend entries present
-    expect(wrapper.text()).toContain('EA1')
-    expect(wrapper.text()).toContain('GA')
-    expect(wrapper.text()).toContain('--')
-    // Total shown in center overlay
-    expect(wrapper.text()).toContain('10')
-    expect(wrapper.text()).toContain('total')
-    // Semantic <dl> legend markup
-    expect(wrapper.find('dl').exists()).toBe(true)
-    expect(wrapper.findAll('dt')).toHaveLength(3)
-    expect(wrapper.findAll('dd')).toHaveLength(3)
-  })
-
-  it('shows empty-state message when releaseDistribution is null', () => {
-    const wrapper = mount(SummaryCards, {
-      props: { summary, releaseDistribution: null }
-    })
-    // Release Version card header is always present
-    expect(wrapper.text()).toContain('Milestone Coverage')
-    // Empty-state message is shown
-    expect(wrapper.text()).toContain('No release version data available')
-    // No <dl> markup rendered
-    expect(wrapper.find('dl').exists()).toBe(false)
-  })
-})
 
 describe('BigRockRow', () => {
   const rock = {
