--- modules/releases/client/plan/components/FeatureHealthRow.vue
@@ -1,383 +1,0 @@
-<script setup>
-import { computed } from 'vue'
-import RiceScoreDisplay from './RiceScoreDisplay.vue'
-import PlanningGateStatus from './PlanningGateStatus.vue'
-import StatusBadge from './StatusBadge.vue'
-import RiskPopover from './RiskPopover.vue'
-
-const props = defineProps({
-  feature: { type: Object, required: true },
-  expanded: { type: Boolean, default: false },
-  canEdit: { type: Boolean, default: false },
-  jiraBaseUrl: { type: String, default: '' },
-  isAdded: { type: Boolean, default: false },
-  showChanges: { type: Boolean, default: true },
-  releasePhaseMode: { type: String, default: 'unknown' }
-})
-
-const emit = defineEmits(['toggle', 'setOverride', 'removeOverride'])
-
-var PLANNING_STATUS_LABELS = {
-  'not-ready': 'Not Ready',
-  'in-planning': 'In Planning',
-  'ready-for-execution': 'Ready'
-}
-
-var PLANNING_STATUS_CLASSES = {
-  'not-ready': 'bg-red-100 dark:bg-red-500/20 text-red-700 dark:text-red-400',
-  'in-planning': 'bg-yellow-100 dark:bg-yellow-500/20 text-yellow-700 dark:text-yellow-400',
-  'ready-for-execution': 'bg-green-100 dark:bg-green-500/20 text-green-700 dark:text-green-400'
-}
-
-var planningLabel = computed(function() {
-  return PLANNING_STATUS_LABELS[props.feature.planningStatus] || '-'
-})
-
-var planningBadgeClass = computed(function() {
-  return PLANNING_STATUS_CLASSES[props.feature.planningStatus] || ''
-})
-
-var riskFlags = computed(function() {
-  if (!props.feature.risk || !props.feature.risk.flags) return []
-  return props.feature.risk.flags
-})
-
-var riskLevel = computed(function() {
-  if (!props.feature.risk) return 'green'
-  return props.feature.risk.level || 'green'
-})
-
-var riskOverride = computed(function() {
-  if (!props.feature.risk) return null
-  return props.feature.risk.override || null
-})
-
-var effectiveRisk = computed(function() {
-  if (riskOverride.value) return riskOverride.value.riskOverride || riskLevel.value
-  return riskLevel.value
-})
-
-var priorityScoreClass = computed(function() {
-  var score = props.feature.priorityScore
-  if (score >= 70) return 'bg-green-100 dark:bg-green-500/20 text-green-700 dark:text-green-400'
-  if (score >= 40) return 'bg-yellow-100 dark:bg-yellow-500/20 text-yellow-700 dark:text-yellow-400'
-  return 'bg-red-100 dark:bg-red-500/20 text-red-700 dark:text-red-400'
-})
-
-var featureUrl = computed(function() {
-  if (props.feature.jiraUrl) return props.feature.jiraUrl
-  if (props.jiraBaseUrl && props.feature.key) return props.jiraBaseUrl + '/' + props.feature.key
-  return ''
-})
-
-function handleToggle() {
-  emit('toggle', props.feature.key)
-}
-
-var isPlanningMode = computed(function() {
-  return props.releasePhaseMode === 'planning'
-})
-
-var planningChecksPct = computed(function() {
-  var pc = props.feature.planningChecks
-  if (!pc || !pc.totalCount) return 0
-  return Math.round((pc.passedCount / pc.totalCount) * 100)
-})
-
-var expandedColspan = computed(function() {
-  return isPlanningMode.value ? 12 : 11
-})
-
-var flagSeverityClass = {
-  high: 'bg-red-50 dark:bg-red-500/10 text-red-700 dark:text-red-400 border-red-200 dark:border-red-500/30',
-  medium: 'bg-yellow-50 dark:bg-yellow-500/10 text-yellow-700 dark:text-yellow-400 border-yellow-200 dark:border-yellow-500/30'
-}
-</script>
-
-<template>
-  <!-- Main row -->
-  <tr
-    class="hover:bg-gray-50 dark:hover:bg-gray-700/50 cursor-pointer"
-    :class="isAdded && showChanges ? 'border-l-4 border-l-green-500 bg-green-50/50 dark:bg-green-500/5' : ''"
-    @click="handleToggle"
-  >
-    <!-- Expand toggle -->
-    <td
-      class="px-2 py-2 border border-gray-300 dark:border-gray-600 w-8 text-center"
-      role="button"
-      tabindex="0"
-      :aria-expanded="expanded"
-      :aria-label="'Expand details for ' + feature.key"
-      @keydown.enter.prevent="handleToggle"
-      @keydown.space.prevent="handleToggle"
-    >
-      <svg
-        class="w-3.5 h-3.5 text-gray-400 transition-transform inline-block"
-        :class="expanded ? 'rotate-90' : ''"
-        fill="none" stroke="currentColor" viewBox="0 0 24 24"
-        aria-hidden="true"
-      >
-        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 5l7 7-7 7" />
-      </svg>
-    </td>
-    <!-- Feature key -->
-    <td class="px-3 py-2 border border-gray-300 dark:border-gray-600">
-      <div class="flex items-center gap-1">
-        <a
-          v-if="featureUrl"
-          :href="featureUrl"
-          target="_blank"
-          rel="noopener"
-          class="text-primary-600 dark:text-blue-400 font-mono text-xs hover:underline"
-          @click.stop
-        >{{ feature.key }}</a>
-        <span v-else class="font-mono text-xs text-gray-700 dark:text-gray-300">{{ feature.key }}</span>
-        <a
-          :href="'#/releases/feature-detail?key=' + encodeURIComponent(feature.key) + '&from=plan'"
-          class="text-gray-400 hover:text-primary-600 dark:hover:text-blue-400"
-          title="View feature detail"
-          @click.stop
-        >
-          <svg class="w-3 h-3" fill="none" stroke="currentColor" viewBox="0 0 24 24" aria-hidden="true">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 6H6a2 2 0 00-2 2v10a2 2 0 002 2h10a2 2 0 002-2v-4M14 4h6m0 0v6m0-6L10 14" />
-          </svg>
-        </a>
-        <span v-if="feature.bigRock && feature.bigRock.includes(', ')" class="ml-1 text-[9px] text-indigo-600 dark:text-indigo-400" :title="'Shared across: ' + feature.bigRock">S</span>
-      </div>
-    </td>
-    <!-- Summary -->
-    <td class="px-3 py-2 text-gray-900 dark:text-gray-100 max-w-[300px] border border-gray-300 dark:border-gray-600 text-xs">
-      <span class="line-clamp-2" :title="feature.summary">{{ feature.summary }}</span>
-    </td>
-    <!-- Status -->
-    <td class="px-3 py-2 border border-gray-300 dark:border-gray-600">
-      <StatusBadge :status="feature.status" />
-    </td>
-    <!-- Planning Checks (planning mode only) -->
-    <td v-if="isPlanningMode" class="px-3 py-2 border border-gray-300 dark:border-gray-600">
-      <template v-if="feature.planningChecks">
-        <div class="flex items-center gap-1.5">
-          <span class="text-xs font-semibold" :class="feature.planningChecks.hasHardBlockers ? 'text-red-600 dark:text-red-400' : 'text-green-600 dark:text-green-400'">
-            {{ feature.planningChecks.passedCount }}/{{ feature.planningChecks.totalCount }}
-          </span>
-          <div class="w-16 h-1.5 bg-gray-200 dark:bg-gray-700 rounded-full overflow-hidden">
-            <div
-              class="h-full rounded-full transition-all"
-              :class="feature.planningChecks.hasHardBlockers ? 'bg-red-500' : 'bg-green-500'"
-              :style="{ width: planningChecksPct + '%' }"
-              role="progressbar"
-              :aria-valuenow="feature.planningChecks.passedCount"
-              :aria-valuemin="0"
-              :aria-valuemax="feature.planningChecks.totalCount"
-              :aria-label="'Planning checks for ' + feature.key"
-            />
-          </div>
-          <span v-if="feature.planningChecks.hasHardBlockers" class="inline-block px-1 py-0.5 rounded text-[9px] font-semibold bg-red-100 dark:bg-red-500/20 text-red-700 dark:text-red-400">
-            {{ feature.planningChecks.hardBlockersFailed.length }}
-          </span>
-        </div>
-      </template>
-      <span v-else class="text-gray-400 dark:text-gray-600 text-xs">--</span>
-    </td>
-    <!-- Health (Risk + Planning Status) -->
-    <td class="px-3 py-2 border border-gray-300 dark:border-gray-600">
-      <div class="flex items-center gap-1.5">
-        <RiskPopover
-          :level="riskLevel"
-          :flags="riskFlags"
-          :flagCount="riskFlags.length"
-          :override="riskOverride"
-          :planningStatus="feature.planningStatus"
-          :fpdor="feature.fpdor"
-          variant="full"
-        >
-          <span
-            class="w-2.5 h-2.5 rounded-full flex-shrink-0"
-            :class="{
-              'cursor-help': riskFlags.length > 0 || riskOverride,
-              'bg-green-500': effectiveRisk === 'green',
-              'bg-yellow-500': effectiveRisk === 'yellow',
-              'bg-red-500': effectiveRisk === 'red'
-            }"
-            role="img"
-            :aria-label="'Risk level: ' + effectiveRisk"
-          ></span>
-        </RiskPopover>
-        <span
-          v-if="feature.planningStatus"
-          class="inline-block px-1.5 py-0.5 rounded text-[10px] font-semibold"
-          :class="planningBadgeClass"
-        >{{ planningLabel }}</span>
-      </div>
-    </td>
-    <!-- Priority -->
-    <td class="px-3 py-2 border border-gray-300 dark:border-gray-600 text-center relative group/priority">
-      <span
-        v-if="feature.priorityScore != null"
-        class="inline-block px-1.5 py-0.5 rounded text-[10px] font-bold"
-        :class="priorityScoreClass"
-      >{{ feature.priorityScore }}</span>
-      <span v-else class="text-gray-400 dark:text-gray-600 text-xs">-</span>
-      <div v-if="feature.priorityBreakdown"
-           class="hidden group-hover/priority:block absolute z-50 bottom-full left-1/2 -translate-x-1/2 mb-1 w-44 bg-gray-900 text-white text-[10px] rounded-lg p-2 shadow-lg pointer-events-none">
-        <div class="font-semibold mb-1">Priority Breakdown</div>
-        <div class="space-y-0.5">
-          <div class="flex justify-between"><span>RICE (30w)</span><span>{{ feature.priorityBreakdown.rice }}%</span></div>
-          <div class="flex justify-between"><span>Big Rock (30w)</span><span>{{ feature.priorityBreakdown.bigRock }}%</span></div>
-          <div class="flex justify-between"><span>Priority (25w)</span><span>{{ feature.priorityBreakdown.priority }}%</span></div>
-          <div class="flex justify-between"><span>Complexity (15w)</span><span>{{ feature.priorityBreakdown.complexity }}%</span></div>
-        </div>
-      </div>
-    </td>
-    <!-- RICE -->
-    <td class="px-3 py-2 border border-gray-300 dark:border-gray-600 text-center">
-      <RiceScoreDisplay :rice="feature.rice" />
-    </td>
-    <!-- Component -->
-    <td class="px-3 py-2 text-xs text-gray-600 dark:text-gray-400 border border-gray-300 dark:border-gray-600">{{ feature.components || '-' }}</td>
-    <!-- Owner -->
-    <td class="px-3 py-2 text-xs text-gray-600 dark:text-gray-400 border border-gray-300 dark:border-gray-600">{{ feature.deliveryOwner || '-' }}</td>
-    <!-- Fix Version -->
-    <td class="px-3 py-2 text-xs text-gray-600 dark:text-gray-400 border border-gray-300 dark:border-gray-600">{{ feature.fixVersions || '-' }}</td>
-    <!-- Target Release -->
-    <td class="px-3 py-2 text-xs text-gray-600 dark:text-gray-400 border border-gray-300 dark:border-gray-600">{{ feature.targetRelease || '-' }}</td>
-  </tr>
-
-  <!-- Expanded detail row -->
-  <tr v-if="expanded">
-    <td :colspan="expandedColspan" class="border border-gray-300 dark:border-gray-600 bg-gray-50 dark:bg-gray-800/50 p-0">
-      <div class="p-4 space-y-4">
-        <div class="grid grid-cols-1 lg:grid-cols-2 gap-4">
-          <!-- Planning Gate Status -->
-          <div class="bg-white dark:bg-gray-800 rounded-lg border border-gray-200 dark:border-gray-700 p-3">
-            <PlanningGateStatus
-              :planningStatus="feature.planningStatus"
-              :fpdor="feature.fpdor"
-            />
-          </div>
-
-          <!-- Risk Flags & Details -->
-          <div class="space-y-3">
-            <!-- Commitment Status -->
-            <div v-if="isAdded && showChanges" class="bg-green-50 dark:bg-green-500/10 rounded-lg border border-green-200 dark:border-green-500/30 p-3">
-              <div class="text-xs font-semibold text-green-700 dark:text-green-400 mb-1">Added After Commitment</div>
-              <div class="text-xs text-green-600 dark:text-green-400/80">
-                This feature was not in the committed list when the snapshot was taken.
-              </div>
-            </div>
-            <!-- Risk flags -->
-            <div v-if="riskFlags.length > 0" class="bg-white dark:bg-gray-800 rounded-lg border border-gray-200 dark:border-gray-700 p-3">
-              <div class="text-xs font-semibold text-gray-700 dark:text-gray-300 mb-2">Risk Flags</div>
-              <div class="space-y-1.5">
-                <div
-                  v-for="(flag, idx) in riskFlags"
-                  :key="idx"
-                  class="flex items-start gap-2 text-xs px-2 py-1.5 rounded border"
-                  :class="flagSeverityClass[flag.severity] || flagSeverityClass.medium"
-                >
-                  <svg
-                    class="w-3.5 h-3.5 flex-shrink-0 mt-0.5"
-                    fill="none" stroke="currentColor" viewBox="0 0 24 24"
-                    aria-hidden="true"
-                  >
-                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-2.5L13.732 4c-.77-.833-1.964-.833-2.732 0L4.082 16.5c-.77.833.192 2.5 1.732 2.5z" />
-                  </svg>
-                  <div>
-                    <span class="font-semibold">{{ flag.category }}</span>
-                    <span class="ml-1">{{ flag.message }}</span>
-                  </div>
-                </div>
-              </div>
-            </div>
-
-            <!-- Override section -->
-            <div v-if="riskOverride" class="bg-white dark:bg-gray-800 rounded-lg border border-gray-200 dark:border-gray-700 p-3">
-              <div class="text-xs font-semibold text-gray-700 dark:text-gray-300 mb-1">Risk Override</div>
-              <div class="text-xs text-gray-600 dark:text-gray-400">
-                <span class="font-medium">Level:</span> {{ riskOverride.riskOverride }}
-              </div>
-              <div class="text-xs text-gray-600 dark:text-gray-400">
-                <span class="font-medium">Reason:</span> {{ riskOverride.reason }}
-              </div>
-              <div class="text-[10px] text-gray-400 dark:text-gray-500 mt-1">
-                Set by {{ riskOverride.updatedBy }} on {{ riskOverride.updatedAt ? new Date(riskOverride.updatedAt).toLocaleDateString() : '' }}
-              </div>
-              <button
-                v-if="canEdit"
-                @click.stop="$emit('removeOverride', feature.key)"
-                class="mt-2 text-[10px] text-red-600 dark:text-red-400 hover:underline"
-              >Remove override</button>
-            </div>
-
-            <!-- Feature metadata -->
-            <div class="bg-white dark:bg-gray-800 rounded-lg border border-gray-200 dark:border-gray-700 p-3">
-              <div class="text-xs font-semibold text-gray-700 dark:text-gray-300 mb-2">Feature Details</div>
-              <div class="grid grid-cols-2 gap-x-4 gap-y-1 text-xs">
-                <div>
-                  <span class="text-gray-500 dark:text-gray-400">PM:</span>
-                  <span class="ml-1 text-gray-900 dark:text-gray-100">{{ feature.pm || '-' }}</span>
-                </div>
-                <div>
-                  <span class="text-gray-500 dark:text-gray-400">Owner:</span>
-                  <span class="ml-1 text-gray-900 dark:text-gray-100">{{ feature.deliveryOwner || '-' }}</span>
-                </div>
-                <div>
-                  <span class="text-gray-500 dark:text-gray-400">Epics:</span>
-                  <span class="ml-1 text-gray-900 dark:text-gray-100">{{ feature.epicCount != null ? feature.epicCount : '-' }}</span>
-                </div>
-                <div>
-                  <span class="text-gray-500 dark:text-gray-400">Issues:</span>
-                  <span class="ml-1 text-gray-900 dark:text-gray-100">{{ feature.issueCount != null ? feature.issueCount : '-' }}</span>
-                </div>
-                <div>
-                  <span class="text-gray-500 dark:text-gray-400">Completion:</span>
-                  <span class="ml-1 text-gray-900 dark:text-gray-100">{{ feature.completionPct != null ? feature.completionPct + '%' : '-' }}</span>
-                </div>
-                <div>
-                  <span class="text-gray-500 dark:text-gray-400">Blockers:</span>
-                  <span
-                    class="ml-1"
-                    :class="feature.blockerCount > 0 ? 'text-red-600 dark:text-red-400 font-semibold' : 'text-gray-900 dark:text-gray-100'"
-                  >{{ feature.blockerCount != null ? feature.blockerCount : '-' }}</span>
-                </div>
-                <div v-if="feature.bigRock">
-                  <span class="text-gray-500 dark:text-gray-400">Big Rock:</span>
-                  <template v-for="(rock, idx) in feature.bigRock.split(', ')" :key="rock">
-                    <span v-if="idx > 0" class="text-gray-400">,</span>
-                    <a
-                      :href="'#/releases/plan?bigRock=' + encodeURIComponent(rock)"
-                      class="ml-1 text-primary-600 dark:text-blue-400 hover:underline"
-                      @click.stop
-                    >{{ rock }}</a>
-                  </template>
-                </div>
-                <div v-if="feature.tier">
-                  <span class="text-gray-500 dark:text-gray-400">Tier:</span>
-                  <span class="ml-1 text-gray-900 dark:text-gray-100">{{ feature.tier }}</span>
-                </div>
-                <div v-if="feature.tshirtSize">
-                  <span class="text-gray-500 dark:text-gray-400">Size:</span>
-                  <span class="ml-1 text-gray-900 dark:text-gray-100">{{ feature.tshirtSize }}</span>
-                </div>
-              </div>
-            </div>
-
-            <!-- Version History -->
-            <div v-if="feature.versionHistory && feature.versionHistory.length > 0"
-                 class="bg-white dark:bg-gray-800 rounded-lg border border-gray-200 dark:border-gray-700 p-3">
-              <div class="text-xs font-semibold text-gray-700 dark:text-gray-300 mb-2">Version History</div>
-              <div class="space-y-1">
-                <div v-for="(change, idx) in feature.versionHistory" :key="idx" class="text-xs text-gray-600 dark:text-gray-400">
-                  <span class="text-gray-400">{{ change.date ? new Date(change.date).toLocaleDateString() : '' }}:</span>
-                  <span v-if="change.from" class="line-through text-red-500 ml-1">{{ change.from }}</span>
-                  <span class="mx-1">&rarr;</span>
-                  <span class="text-green-600">{{ change.to }}</span>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </td>
-  </tr>
-</template>
