--- modules/releases/client/plan/components/FeaturesTable.vue
@@ -1,199 +1,0 @@
-<script setup>
-import StatusBadge from './StatusBadge.vue'
-import RiskBadge from './RiskBadge.vue'
-import RiskPopover from './RiskPopover.vue'
-import TierSeparator from './TierSeparator.vue'
-import { computed, inject } from 'vue'
-import { useModuleLink } from '@shared/client/composables/useModuleLink.js'
-import { PRIORITY_STYLES } from '../utils/constants'
-
-const moduleNav = inject('moduleNav')
-const aiImpactAvailable = computed(() => moduleNav?.isModuleAvailable?.('ai-impact') ?? false)
-const { linkTo } = useModuleLink()
-
-const props = defineProps({
-  features: { type: Array, default: () => [] },
-  bigRocks: { type: Array, default: () => [] },
-  jiraBaseUrl: { type: String, default: '' },
-  summary: { type: Object, default: null },
-  healthByKey: { type: Object, default: () => ({}) }
-})
-
-const hasHealth = computed(function() {
-  return Object.keys(props.healthByKey).length > 0
-})
-
-const COL_COUNT = computed(function() {
-  return hasHealth.value ? 13 : 11
-})
-
-const tierCounts = computed(() => {
-  const counts = {}
-  for (const f of props.features) {
-    const tier = f.tier || 1
-    counts[tier] = (counts[tier] || 0) + 1
-  }
-  return counts
-})
-
-const groupedFeatures = computed(() => {
-  const groups = []
-  let currentTier = null
-
-  for (const f of props.features) {
-    const tier = f.tier || 1
-    if (tier !== currentTier) {
-      currentTier = tier
-      groups.push({ type: 'separator', tier, count: tierCounts.value[tier] || 0 })
-    }
-    groups.push({ type: 'feature', data: f })
-  }
-
-  return groups
-})
-</script>
-
-<template>
-  <div class="bg-white dark:bg-gray-800 rounded-lg border border-gray-300 dark:border-gray-600 overflow-hidden">
-    <div class="overflow-x-auto">
-      <table class="w-full text-sm border-collapse">
-        <caption class="sr-only">Features grouped by tier</caption>
-        <thead>
-          <tr>
-            <th scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">Big Rock</th>
-            <th scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">Feature</th>
-            <th scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">Status</th>
-            <th v-if="hasHealth" scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">Risk</th>
-            <th v-if="hasHealth" scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">Planning</th>
-            <th scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">Priority</th>
-            <th scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">Phase</th>
-            <th scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">Title</th>
-            <th scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">Components</th>
-            <th scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">Target Release</th>
-            <th scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">Delivery Owner</th>
-            <th scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">RFE</th>
-            <th scope="col" class="px-3 py-2 text-left text-gray-700 dark:text-gray-200 font-semibold uppercase text-xs tracking-wide border border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-900/80">Fix Version</th>
-          </tr>
-        </thead>
-        <tbody>
-          <template v-for="(item, idx) in groupedFeatures" :key="idx">
-            <TierSeparator
-              v-if="item.type === 'separator'"
-              :tier="item.tier"
-              :count="item.count"
-              :colspan="COL_COUNT"
-            />
-            <tr
-              v-else
-              class="hover:bg-gray-50 dark:hover:bg-gray-700/50"
-            >
-              <td class="px-3 py-2 text-xs text-gray-600 dark:text-gray-400 max-w-[120px] truncate border border-gray-300 dark:border-gray-600">{{ item.data.bigRock || '-' }}</td>
-              <td class="px-3 py-2 border border-gray-300 dark:border-gray-600">
-                <div class="flex items-center gap-1">
-                  <a
-                    :href="'#/releases/feature-detail?key=' + encodeURIComponent(item.data.issueKey) + '&from=plan'"
-                    class="text-primary-600 dark:text-blue-400 font-mono text-xs hover:underline"
-                  >{{ item.data.issueKey }}</a>
-                  <a
-                    :href="item.data.jiraUrl || `${jiraBaseUrl}/${item.data.issueKey}`"
-                    target="_blank"
-                    rel="noopener"
-                    class="text-gray-400 hover:text-primary-600 dark:hover:text-blue-400 shrink-0"
-                    title="Open in Jira"
-                  >
-                    <svg class="w-3 h-3" fill="none" stroke="currentColor" viewBox="0 0 24 24" aria-hidden="true">
-                      <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 6H6a2 2 0 00-2 2v10a2 2 0 002 2h10a2 2 0 002-2v-4M14 4h6m0 0v6m0-6L10 14" />
-                    </svg>
-                  </a>
-                </div>
-              </td>
-              <td class="px-3 py-2 border border-gray-300 dark:border-gray-600"><StatusBadge :status="item.data.status" /></td>
-              <td v-if="hasHealth" class="px-3 py-2 border border-gray-300 dark:border-gray-600">
-                <RiskPopover
-                  v-if="healthByKey[item.data.issueKey]"
-                  :level="healthByKey[item.data.issueKey].risk ? healthByKey[item.data.issueKey].risk.level : 'green'"
-                  :flags="healthByKey[item.data.issueKey].risk ? healthByKey[item.data.issueKey].risk.flags : []"
-                  :flagCount="healthByKey[item.data.issueKey].risk ? healthByKey[item.data.issueKey].risk.flags.length : 0"
-                  :override="healthByKey[item.data.issueKey].risk ? healthByKey[item.data.issueKey].risk.override : null"
-                  :planningStatus="healthByKey[item.data.issueKey].planningStatus || ''"
-                  variant="full"
-                >
-                  <RiskBadge
-                    :level="healthByKey[item.data.issueKey].risk ? healthByKey[item.data.issueKey].risk.level : 'green'"
-                    :flagCount="healthByKey[item.data.issueKey].risk ? healthByKey[item.data.issueKey].risk.flags.length : 0"
-                    :flags="healthByKey[item.data.issueKey].risk ? healthByKey[item.data.issueKey].risk.flags : []"
-                    :override="healthByKey[item.data.issueKey].risk ? healthByKey[item.data.issueKey].risk.override : null"
-                  />
-                </RiskPopover>
-                <span v-else class="text-gray-400 dark:text-gray-600 text-xs">-</span>
-              </td>
-              <td v-if="hasHealth" class="px-3 py-2 border border-gray-300 dark:border-gray-600 text-xs text-center">
-                <span
-                  v-if="healthByKey[item.data.issueKey] && healthByKey[item.data.issueKey].planningStatus"
-                  class="inline-block px-1.5 py-0.5 rounded text-[10px] font-semibold"
-                  :class="{
-                    'bg-red-100 dark:bg-red-500/20 text-red-700 dark:text-red-400': healthByKey[item.data.issueKey].planningStatus === 'not-ready',
-                    'bg-yellow-100 dark:bg-yellow-500/20 text-yellow-700 dark:text-yellow-400': healthByKey[item.data.issueKey].planningStatus === 'in-planning',
-                    'bg-green-100 dark:bg-green-500/20 text-green-700 dark:text-green-400': healthByKey[item.data.issueKey].planningStatus === 'ready-for-execution'
-                  }"
-                >{{ { 'not-ready': 'Not Ready', 'in-planning': 'In Planning', 'ready-for-execution': 'Ready' }[healthByKey[item.data.issueKey].planningStatus] || '-' }}</span>
-                <span v-else class="text-gray-400 dark:text-gray-600">-</span>
-              </td>
-              <td class="px-3 py-2 border border-gray-300 dark:border-gray-600">
-                <span
-                  v-if="item.data.priority"
-                  class="inline-block px-1.5 py-0.5 rounded text-[10px] font-semibold"
-                  :class="PRIORITY_STYLES[item.data.priority] || PRIORITY_STYLES['Normal']"
-                >{{ item.data.priority }}</span>
-              </td>
-              <td class="px-3 py-2 border border-gray-300 dark:border-gray-600">
-                <span
-                  v-if="item.data.phase"
-                  class="inline-block px-1.5 py-0.5 rounded text-[10px] font-semibold"
-                  :class="{
-                    'bg-amber-100 text-amber-800 dark:bg-amber-900/30 dark:text-amber-400': item.data.phase === 'TP',
-                    'bg-blue-100 text-blue-800 dark:bg-blue-900/30 dark:text-blue-400': item.data.phase === 'DP',
-                    'bg-green-100 text-green-800 dark:bg-green-900/30 dark:text-green-400': item.data.phase === 'GA'
-                  }"
-                >{{ item.data.phase }}</span>
-                <span v-else class="text-gray-400 dark:text-gray-600 text-xs">-</span>
-              </td>
-              <td class="px-3 py-2 text-gray-900 dark:text-gray-100 max-w-xs truncate border border-gray-300 dark:border-gray-600">{{ item.data.summary }}</td>
-              <td class="px-3 py-2 text-xs text-gray-600 dark:text-gray-400 border border-gray-300 dark:border-gray-600">{{ item.data.components }}</td>
-              <td class="px-3 py-2 text-xs text-gray-600 dark:text-gray-400 border border-gray-300 dark:border-gray-600">{{ item.data.targetRelease }}</td>
-              <td class="px-3 py-2 text-xs text-gray-600 dark:text-gray-400 border border-gray-300 dark:border-gray-600">{{ item.data.deliveryOwner || '-' }}</td>
-              <td class="px-3 py-2 border border-gray-300 dark:border-gray-600">
-                <div v-if="item.data.rfe" class="flex items-center gap-1">
-                  <a
-                    v-if="aiImpactAvailable"
-                    :href="linkTo('ai-impact', 'rfe-review', { select: item.data.rfe })"
-                    class="text-primary-600 dark:text-blue-400 font-mono text-xs hover:underline"
-                  >{{ item.data.rfe }}</a>
-                  <span v-else class="font-mono text-xs text-gray-900 dark:text-gray-100">{{ item.data.rfe }}</span>
-                  <a
-                    :href="`${jiraBaseUrl}/${item.data.rfe}`"
-                    target="_blank"
-                    rel="noopener"
-                    class="text-gray-400 hover:text-primary-600 dark:hover:text-blue-400 shrink-0"
-                    title="Open in Jira"
-                  >
-                    <svg class="w-3 h-3" fill="none" stroke="currentColor" viewBox="0 0 24 24" aria-hidden="true">
-                      <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 6H6a2 2 0 00-2 2v10a2 2 0 002 2h10a2 2 0 002-2v-4M14 4h6m0 0v6m0-6L10 14" />
-                    </svg>
-                  </a>
-                </div>
-                <span v-else class="text-gray-400 dark:text-gray-600 text-xs">-</span>
-              </td>
-              <td class="px-3 py-2 text-xs text-gray-600 dark:text-gray-400 border border-gray-300 dark:border-gray-600">{{ item.data.fixVersion || '-' }}</td>
-            </tr>
-          </template>
-          <tr v-if="!features || features.length === 0">
-            <td :colspan="COL_COUNT" class="px-3 py-8 text-center text-gray-500 border border-gray-300 dark:border-gray-600">
-              No features found matching the current filters.
-            </td>
-          </tr>
-        </tbody>
-      </table>
-    </div>
-  </div>
-</template>
