--- modules/releases/client/plan/utils/constants.js
@@ -1,7 +1,0 @@
-export const PRIORITY_STYLES = {
-  'Blocker': 'bg-red-100 dark:bg-red-500/20 text-red-700 dark:text-red-400',
-  'Critical': 'bg-orange-100 dark:bg-orange-500/20 text-orange-700 dark:text-orange-400',
-  'Major': 'bg-yellow-100 dark:bg-yellow-500/20 text-yellow-700 dark:text-yellow-400',
-  'Normal': 'bg-blue-100 dark:bg-blue-500/20 text-blue-700 dark:text-blue-400',
-  'Minor': 'bg-gray-100 dark:bg-gray-500/20 text-gray-600 dark:text-gray-400'
-}
